requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
idna==3.4
    # via requests
pydantic[dotenv]==1.10.2
    # via -r requirements.in
python-dotenv==0.21.0
    # via pydantic
requests==2.28.1
    # via -r requirements.in
typing-extensions==4.4.0
    # via pydantic
urllib3==1.26.12
    # via requests
websocket-client==1.4.1
    # via -r requirements.in
